Module aws_sdk_signer::error

Modules

access_denied_exception

See AccessDeniedException

bad_request_exception

See BadRequestException

conflict_exception

See ConflictException

internal_service_error_exception

See InternalServiceErrorException

not_found_exception

See NotFoundException

resource_not_found_exception

See ResourceNotFoundException

service_limit_exceeded_exception

See ServiceLimitExceededException

throttling_exception

See ThrottlingException

too_many_requests_exception

See TooManyRequestsException

validation_exception

See ValidationException

Structs

AccessDeniedException

You do not have sufficient access to perform this action.

AddProfilePermissionError

BadRequestException

The request contains invalid parameters for the ARN or tags. This exception also

CancelSigningProfileError

ConflictException

The resource encountered a conflicting state.

DescribeSigningJobError

GetSigningPlatformError

GetSigningProfileError

InternalServiceErrorException

An internal error occurred.

ListProfilePermissionsError

ListSigningJobsError

ListSigningPlatformsError

ListSigningProfilesError

ListTagsForResourceError

NotFoundException

The signing profile was not found.

PutSigningProfileError

RemoveProfilePermissionError

ResourceNotFoundException

A specified resource could not be found.

RevokeSignatureError

RevokeSigningProfileError

ServiceLimitExceededException

The client is making a request that exceeds service limits.

StartSigningJobError

TagResourceError

ThrottlingException

The request was denied due to request throttling.

TooManyRequestsException

The allowed number of job-signing requests has been exceeded.

UntagResourceError

ValidationException

You signing certificate could not be validated.

Enums

AddProfilePermissionErrorKind

CancelSigningProfileErrorKind

DescribeSigningJobErrorKind

GetSigningPlatformErrorKind

GetSigningProfileErrorKind

ListProfilePermissionsErrorKind

ListSigningJobsErrorKind

ListSigningPlatformsErrorKind

ListSigningProfilesErrorKind

ListTagsForResourceErrorKind

PutSigningProfileErrorKind

RemoveProfilePermissionErrorKind

RevokeSignatureErrorKind

RevokeSigningProfileErrorKind

StartSigningJobErrorKind

TagResourceErrorKind

UntagResourceErrorKind