Module aws_sdk_signer::model

Modules

destination

See Destination

encryption_algorithm_options

See EncryptionAlgorithmOptions

hash_algorithm_options

See HashAlgorithmOptions

permission

See Permission

s3_destination

See S3Destination

s3_signed_object

See S3SignedObject

s3_source

See S3Source

signature_validity_period

See SignatureValidityPeriod

signed_object

See SignedObject

signing_configuration

See SigningConfiguration

signing_configuration_overrides

See SigningConfigurationOverrides

signing_image_format

See SigningImageFormat

signing_job

See SigningJob

signing_job_revocation_record

See SigningJobRevocationRecord

signing_material

See SigningMaterial

signing_platform

See SigningPlatform

signing_platform_overrides

See SigningPlatformOverrides

signing_profile

See SigningProfile

signing_profile_revocation_record

See SigningProfileRevocationRecord

source

See Source

Structs

Destination

Points to an S3Destination object that contains information about your S3

EncryptionAlgorithmOptions

The encryption algorithm options that are available to a code signing job.

HashAlgorithmOptions

The hash algorithms that are available to a code signing job.

Permission

A cross-account permission for a signing profile.

S3Destination

The name and prefix of the S3 bucket where code signing saves your signed objects.

S3SignedObject

The S3 bucket name and key where code signing saved your signed code image.

S3Source

Information about the S3 bucket where you saved your unsigned code.

SignatureValidityPeriod

The validity period for a signing job.

SignedObject

Points to an S3SignedObject object that contains information about your

SigningConfiguration

The configuration of a code signing operation.

SigningConfigurationOverrides

A signing configuration that overrides the default encryption or hash algorithm of a

SigningImageFormat

The image format of a code signing platform or profile.

SigningJob

Contains information about a signing job.

SigningJobRevocationRecord

Revocation information for a signing job.

SigningMaterial

The ACM certificate that is used to sign your code.

SigningPlatform

Contains information about the signing configurations and parameters that are used to

SigningPlatformOverrides

Any overrides that are applied to the signing configuration of a code signing

SigningProfile

Contains information about the ACM certificates and code signing configuration parameters

SigningProfileRevocationRecord

Revocation information for a signing profile.

Source

An S3Source object that contains information about the S3 bucket where

Enums

Category

EncryptionAlgorithm

HashAlgorithm

ImageFormat

SigningProfileStatus

SigningStatus

ValidityType