Module aws_sdk_signer::operation

Structs

AddProfilePermission

Adds cross-account permissions to a signing profile.

CancelSigningProfile

Changes the state of an ACTIVE signing profile to CANCELED.

DescribeSigningJob

Returns information about a specific code signing job. You specify the job by using

GetSigningPlatform

Returns information on a specific signing platform.

GetSigningProfile

Returns information on a specific signing profile.

ListProfilePermissions

Lists the cross-account permissions associated with a signing profile.

ListSigningJobs

Lists all your signing jobs. You can use the maxResults parameter to

ListSigningPlatforms

Lists all signing platforms available in code signing that match the request parameters. If

ListSigningProfiles

Lists all available signing profiles in your AWS account. Returns only profiles with

ListTagsForResource

Returns a list of the tags associated with a signing profile resource.

PutSigningProfile

Creates a signing profile. A signing profile is a code signing template that can be used to

RemoveProfilePermission

Removes cross-account permissions from a signing profile.

RevokeSignature

Changes the state of a signing job to REVOKED. This indicates that the signature is no

RevokeSigningProfile

Changes the state of a signing profile to REVOKED. This indicates that signatures

StartSigningJob

Initiates a signing job to be performed on the code provided. Signing jobs are

TagResource

Adds one or more tags to a signing profile. Tags are labels that you can use to

UntagResource

Removes one or more tags from a signing profile. To remove the tags, specify a list of