Struct aws_sdk_signer::output::DescribeSigningJobOutput

#[non_exhaustive]
pub struct DescribeSigningJobOutput {

    pub job_id: Option<String>,
    pub source: Option<Source>,
    pub signing_material: Option<SigningMaterial>,
    pub platform_id: Option<String>,
    pub platform_display_name: Option<String>,
    pub profile_name: Option<String>,
    pub profile_version: Option<String>,
    pub overrides: Option<SigningPlatformOverrides>,
    pub signing_parameters: Option<HashMap<String, String>>,
    pub created_at: Option<Instant>,
    pub completed_at: Option<Instant>,
    pub signature_expires_at: Option<Instant>,
    pub requested_by: Option<String>,
    pub status: Option<SigningStatus>,
    pub status_reason: Option<String>,
    pub revocation_record: Option<SigningJobRevocationRecord>,
    pub signed_object: Option<SignedObject>,
    pub job_owner: Option<String>,
    pub job_invoker: Option<String>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

job_id: Option<String>

The ID of the signing job on output.

source: Option<Source>

The object that contains the name of your S3 bucket or your raw code.

signing_material: Option<SigningMaterial>

The Amazon Resource Name (ARN) of your code signing certificate.

platform_id: Option<String>

The microcontroller platform to which your signed code image will be distributed.

platform_display_name: Option<String>

A human-readable name for the signing platform associated with the signing job.

profile_name: Option<String>

The name of the profile that initiated the signing operation.

profile_version: Option<String>

The version of the signing profile used to initiate the signing job.

overrides: Option<SigningPlatformOverrides>

A list of any overrides that were applied to the signing operation.

signing_parameters: Option<HashMap<String, String>>

Map of user-assigned key-value pairs used during signing. These values contain any information that you specified for use in your signing job.

created_at: Option<Instant>

Date and time that the signing job was created.

completed_at: Option<Instant>

Date and time that the signing job was completed.

signature_expires_at: Option<Instant>

Thr expiration timestamp for the signature generated by the signing job.

requested_by: Option<String>

The IAM principal that requested the signing job.

status: Option<SigningStatus>

Status of the signing job.

status_reason: Option<String>

String value that contains the status reason.

revocation_record: Option<SigningJobRevocationRecord>

A revocation record if the signature generated by the signing job has been revoked. Contains a timestamp and the ID of the IAM entity that revoked the signature.

signed_object: Option<SignedObject>

Name of the S3 bucket where the signed code image is saved by code signing.

job_owner: Option<String>

The AWS account ID of the job owner.

job_invoker: Option<String>

The IAM entity that initiated the signing job.

Implementations

impl DescribeSigningJobOutput

pub fn builder() -> Builder

Creates a new builder-style object to manufacture DescribeSigningJobOutput